[REFACTOR] 단계 분리 (계산 / 출력)
